requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.7.0
email-validator>=2.1.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0

# Data ingestion
requests>=2.31.0
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.0
finnhub-python>=2.4.0
newsapi-python>=0.2.7
rapidfuzz>=3.0.0

# ML (Isolation Forest + NetworkX fallback — GNN plugs in later)
scikit-learn>=1.3.0
networkx>=3.0

# Utilities
python-multipart>=0.0.9
